link sword works!
